Parse long integers as longs.
The int and long built-ins are almost identical for most Python code,
except that GMPY refuses to convert mpz values to long integers using
int and one must use long instead. Python 3000 will fix this by having
only one integer type.

+++ viff/config.py
@@ -126,10 +126,10 @@ def load_config(source):
         id = p_unstr(player)
         host = config[player]['host']
         port = int(config[player]['port'])
-        pubkey = tuple(map(int, config[player]['pubkey']))
+        pubkey = tuple(map(long, config[player]['pubkey']))
 
         if 'prss_keys' in config[player]:
-            seckey = tuple(map(int, config[player]['seckey']))
+            seckey = tuple(map(long, config[player]['seckey']))
             keys = {}
             for subset in config[player]['prss_keys']:
                 keys[s_unstr(subset)] = config[player]['prss_keys'][subset]
